ba-ner/requirements.txt
torch>=2.4.0
transformers>=4.48.0
datasets>=3.0.0
accelerate>=1.2.0
peft>=0.14.0
trl>=0.13.0
bitsandbytes>=0.45.0
seqeval>=1.2.2
scikit-learn>=1.5.0
wandb>=0.19.0
pyyaml>=6.0
rich>=13.0
matplotlib>=3.9.0
pandas>=2.2.0
flash-attn>=2.7.0
numpy>=1.26.0
